requirements-rag.txt
# =============================================
# Sarathi — RAG Pipeline Requirements (rag-venv)
# =============================================
# Python 3.10+ recommended
#
# Setup:
#   python -m venv rag-venv
#   rag-venv\Scripts\activate           (Windows)
#   source rag-venv/bin/activate        (Linux/Mac)
#   pip install -r requirements-rag.txt
#
# For CUDA GPU support (recommended for embeddings):
#   pip install torch==2.6.0+cu124 torchvision==0.21.0+cu124 torchaudio==2.6.0+cu124 \
#       --index-url https://download.pytorch.org/whl/cu124
# Then install the rest:
#   pip install -r requirements-rag.txt
# =============================================

# --- Web Framework ---
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# --- LLM ---
google-generativeai>=0.8.0

# --- Embeddings ---
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
sentence-transformers>=3.0.0
transformers>=4.40.0
huggingface_hub>=0.20.0

# --- Vector Store ---
chromadb>=0.5.0

# --- PDF Processing ---
pypdf>=4.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10

# --- PyTorch (CPU fallback — see header for CUDA) ---
torch>=2.6.0
safetensors>=0.4.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
